Alarm id modified
The alarm id was being sent as static but it initialised on app startup every time. Therefore id corresponding to the item was used as AlarmId. This was possible because my app doesnt contain a fab now. It has a quick text. So, the id will always!=-1 and hence a unique Id.